feat: refactor geofence constants and integrate into work location components
