--- frontend/src/community/configurations/constants/geofenceConstants.ts
@@ -1,2 +1,0 @@
-export const MIN_RADIUS = 100;
-export const MAX_RADIUS = 2000;
