+++ frontend/src/community/configurations/constants/workLocationConstants.ts
@@ -1,0 +1,4 @@
+export const WORK_LOCATION_PAGE_SIZE = 4;
+export const WORK_LOCATION_SEARCH_DEBOUNCE_MS = 500;
+export const MIN_RADIUS = 100;
+export const MAX_RADIUS = 2000;
